Split client-preset - fragment masking tests
